fix: add empty check for options
channelId, channelSecret and channelAccessToken should be
present for the bot in order to function properly.

Past behaviour: when a channelSecret is not given, webhook requests
are simply returned 400 Bad Request by the verify function.

After this commit, the constructor will throw an error when any of the
above options are not given.

Haven't written the tests yet.

Signed-off-by: Birkhoff Lee <git@birkhoff.me>